+++ packages/framer-motion/src/gestures/pan/index.ts
@@ -8,7 +8,7 @@ type PanEventHandler = (event: PointerEvent, info: PanInfo) => void
 const asyncHandler =
     (handler?: PanEventHandler) => (event: PointerEvent, info: PanInfo) => {
         if (handler) {
-            frame.postRender(() => handler(event, info))
+            frame.update(() => handler(event, info), false, true)
         }
     }
 
@@ -35,7 +35,7 @@ export class PanGesture extends Feature<Element> {
         return {
             onSessionStart: asyncHandler(onPanSessionStart),
             onStart: asyncHandler(onPanStart),
-            onMove: onPan,
+            onMove: asyncHandler(onPan),
             onEnd: (event: PointerEvent, info: PanInfo) => {
                 delete this.session
                 if (onPanEnd) {